현재 데이터베이스 외래키 잘못해놔서 긴급수정중~~~~~~~~~~~~~~~~~~~~~~으아아앙

+++ src/main/java/com/moveon/server/dto/CommentsRequestDto.java
@@ -16,7 +16,7 @@ public class CommentsRequestDto {
     private String content;
     private int classNum;
     private int num;
-    private int groupId;
+    private Long groupId;
 
     public Comments toComments(){
         return Comments.builder()
